Remove corrupted update of benchmark decoder h264
git-svn-id: https://www.soclib.fr/svn/trunk/soclib@1564 2cf4c1f0-a32b-0410-9068-d80f8be93c97

+++ soclib/module/internal_component/vci_ram/caba/source/src/vci_ram.cpp
@@ -147,8 +147,8 @@ tmpl(void)::genMoore()
 
 tmpl(void)::print_stats(){
     std::cout << name() << std::endl;
-    std::cout << "- READ               = " << m_cpt_read << std::endl;
-    std::cout << "- WRITE              = " << m_cpt_write << std::endl;
+    std::cout << "- READ               = " << (float)m_cpt_read << std::endl;
+    std::cout << "- WRITE              = " << (float)m_cpt_write << std::endl;
 }
 
 }}
